requirements.txt
langgraph>=0.1.18
langchain>=0.2.14
langchain-openai>=0.2.2
langchain-community>=0.2.11
yfinance>=0.2.44
ta>=0.11.0
pandas>=2.2.2
python-dotenv>=1.0.1
